requirements.txt
APIFlask==2.0.1
apispec==6.3.0
Authlib==1.2.1
bidict==0.22.1
black==23.7.0
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.3
Flask==2.3.3
Flask-HTTPAuth==4.8.0
flask-marshmallow==0.15.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.5
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
psycopg2-binary==2.9.7
pycparser==2.21
python-dotenv==1.0.0
python-engineio==4.7.1
python-socketio==5.9.0
requests==2.31.0
simple-websocket==0.10.1
SQLAlchemy==2.0.20
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.4
webargs==8.3.0
Werkzeug==2.3.7
wsproto==1.2.0
zipp==3.16.2
